+++ frontend/src/features/map_planning/hooks/mapEditorHookApi.ts
@@ -4,6 +4,7 @@ import { useTranslation } from 'react-i18next';
 import { getDrawings } from '../api/drawingApi';
 import { getLayers } from '../api/getLayers';
 import { getMap } from '../api/getMap';
+import { getTimelineEvents } from '../api/getTimelineEvents';
 import { getPlantings } from '../api/plantingApi';
 import { getBaseLayerImage } from '../layers/base/api/getBaseLayer';
 import useMapStore from '../store/MapStore';
@@ -15,6 +16,7 @@ const MAP_EDITOR_KEYS = {
     layers: () => [{ ...MAP_EDITOR_KEYS._helpers.all[0], scope: 'layers' }] as const,
     plant: () => [{ ...MAP_EDITOR_KEYS._helpers.all[0], scope: 'plant_layer' }] as const,
     base: () => [{ ...MAP_EDITOR_KEYS._helpers.all[0], scope: 'base_layer' }] as const,
+    timeline: () => [{ ...MAP_EDITOR_KEYS._helpers.all[0], scope: 'timeline' }] as const,
     drawing: () => [{ ...MAP_EDITOR_KEYS._helpers.all[0], scope: 'drawing_layer' }] as const,
   },
   layers: (mapId: number) => [{ ...MAP_EDITOR_KEYS._helpers.layers()[0], mapId }] as const,
@@ -25,9 +27,32 @@ const MAP_EDITOR_KEYS = {
     [{ ...MAP_EDITOR_KEYS._helpers.base()[0], mapId, layerId }] as const,
   drawingLayer: (mapId: number, layerId: number, fetchDate: string) =>
     [{ ...MAP_EDITOR_KEYS._helpers.drawing()[0], mapId, layerId, fetchDate }] as const,
+  timeline: (mapId: number, startDate: string, endDate: string) =>
+    [{ ...MAP_EDITOR_KEYS._helpers.timeline()[0], mapId, startDate, endDate }] as const,
 };
 
 const TEN_MINUTES = 1000 * 60 * 10;
+
+/**
+ * Fetch timeline events for the given map id.
+ */
+export function useGetTimelineEvents(mapId: number, startDate: string, endDate: string) {
+  const queryInfo = useQuery({
+    queryKey: MAP_EDITOR_KEYS.timeline(mapId, startDate, endDate),
+    queryFn: getTimelineEventsQueryFn,
+    refetchOnWindowFocus: false,
+    staleTime: TEN_MINUTES,
+  });
+
+  return queryInfo.data;
+}
+
+function getTimelineEventsQueryFn({
+  queryKey,
+}: QueryFunctionContext<ReturnType<(typeof MAP_EDITOR_KEYS)['timeline']>>) {
+  const { mapId, startDate, endDate } = queryKey[0];
+  return getTimelineEvents(mapId, startDate, endDate);
+}
 
 /**
  * Gets all layers for the given map id.