+++ doc/changelog.md
@@ -10,7 +10,7 @@ Syntax: `- short text describing the change _(Your Name)_`
 
 - needs new migrations
 - needs new scraper data (integer for plant spread and height)
-- _()_
+- pin python package versions for e2e tests #1200 _(4ydan)_
 - Remove variety from table seeds _(Jannis Adamek)_
 - Add timeline API that summarizes plantings #967 _(Jannis Adamek)_
 - Refactor react query error handling _(Daniel Steinkogler)_
@@ -46,6 +46,9 @@ Syntax: `- short text describing the change _(Your Name)_`
 - Migrate from Jest to Vitest, update Vite to v5, update Node to 20, .env should be .env.local _(Paul)_
 - Fix broken .env reading for scraper _(Paul)_
 - Add a dev env variable for react query when developing in offline env (e.g. plane) _(Paul)_
+- Collect events from planting layer and add it to timeline _(Daniel Steinkogler)_
+- _()_
+- _()_
 - Remove error messages in console if a seed was not found _(Moritz)_
 - Add drawing layer functionality including rectangles, ellipses, lines and bezier polygons _(Daniel Steinkogler, Jannis Adamek)_
 - _()_
@@ -62,7 +65,7 @@ Syntax: `- short text describing the change _(Your Name)_`
 - _()_
 - Add key combinations for map geometry _(Daniel Steinkogler)_
 - _()_
-- _()_
+- Add tooltips to show keybindings _(Daniel Steinkogler)_
 - _()_
 - Use arrow keys to navigate between timepicker sliders and fix timeline performance issues _(Daniel Steinkogler)_
 - _()_
@@ -78,7 +81,7 @@ Syntax: `- short text describing the change _(Your Name)_`
 - _()_
 - Added missing error toast messages _(Daniel Steinkogler)_
 - _()_
-- _()_
+- Added keybindings for undo and redo _(Daniel Steinkogler)_
 - E2E: fix planting cursor tests _(4ydan & absurd-turtle)_
 - _()_
 - Added planting notes to planting details _(Daniel Steinkogler)_